moved updateStatus into user route
removed updateUserStatus route and moved updateStatus function to user route
